requirements.txt
Django==5.0.4
python-dotenv==1.0.1
whitenoise==6.6.0
gunicorn==21.2.0
chardet==5.2.0
